+++ CHANGELOG.md
@@ -1,1 +1,2 @@
 - Added "Disable Auto Save Only" function in Advanced Settings (Thanks [LiQiuDGG](https://github.com/NinerBull/AccWideUILayoutSelection/pull/21)!)
+- Added LibDataBroker Support, so you can now add a icon for AWI to Titan Panel, etc.